简单计算-改为Double

+++ 1.变量、运算符、表达式、输入与输出/exercises/acwing_611.java
@@ -5,13 +5,13 @@ public class acwing_611 {
         Scanner sc = new Scanner(System.in);
         int no1, no2;
         int num1, num2;
-        float price1, price2;
+        Double price1, price2;
         no1 = sc.nextInt();
         num1 = sc.nextInt();
-        price1 = sc.nextFloat();
+        price1 = sc.nextDouble();
         no2 = sc.nextInt();
         num2 = sc.nextInt();
-        price2 = sc.nextFloat();
+        price2 = sc.nextDouble();
         System.out.printf("VALOR A PAGAR: R$ %.2f", num1 * price1 + num2 * price2);
         sc.close();
     }
